+++ custom_components/econet300/const.py
@@ -27,6 +27,7 @@ DEVICE_INFO_MODEL = "ecoNET300"
 DEVICE_INFO_CONTROLLER_NAME = "PLUM ecoNET300"
 DEVICE_INFO_MIXER_NAME = "Mixer device"
 DEVICE_INFO_LAMBDA_NAME = "Module Lambda"
+DEVICE_INFO_ECOSTER_NAME = "ecoSTER"
 
 CONF_ENTRY_TITLE = "ecoNET300"
 CONF_ENTRY_DESCRIPTION = "PLUM Econet300"
@@ -72,7 +73,7 @@ API_EDITABLE_PARAMS_LIMITS_DATA = "data"
 ###################################
 #    NUMBER of AVAILABLE MIXERS
 ###################################
-AVAILABLE_NUMBER_OF_MIXERS = 6
+AVAILABLE_NUMBER_OF_MIXERS = 6  # Supports up to 6 mixers (ecoMAX850R2-X has 5)
 MIXER_AVAILABILITY_KEY = "mixerTemp"
 MIXER_SET_AVAILABILITY_KEY = "mixerSetTemp"
 
@@ -107,6 +108,36 @@ SENSOR_MAP_KEY = {
         "Circuit1thermostat",
         "heating_work_state_pump4",
     },
+    # ecoSTER thermostat sensors if moduleEcoSTERSoftVer is not None
+    "ecoSter": {
+        # ecoSTER temperature sensors
+        "ecoSterTemp1",
+        "ecoSterTemp2",
+        "ecoSterTemp3",
+        "ecoSterTemp4",
+        "ecoSterTemp5",
+        "ecoSterTemp6",
+        "ecoSterTemp7",
+        "ecoSterTemp8",
+        # ecoSTER setpoint sensors
+        "ecoSterSetTemp1",
+        "ecoSterSetTemp2",
+        "ecoSterSetTemp3",
+        "ecoSterSetTemp4",
+        "ecoSterSetTemp5",
+        "ecoSterSetTemp6",
+        "ecoSterSetTemp7",
+        "ecoSterSetTemp8",
+        # ecoSTER mode sensors
+        "ecoSterMode1",
+        "ecoSterMode2",
+        "ecoSterMode3",
+        "ecoSterMode4",
+        "ecoSterMode5",
+        "ecoSterMode6",
+        "ecoSterMode7",
+        "ecoSterMode8",
+    },
     "lambda": {
         "lambdaStatus",
         "lambdaSet",
@@ -138,6 +169,13 @@ SENSOR_MAP_KEY = {
         "moduleCSoftVer",
         "moduleLambdaSoftVer",
         "modulePanelSoftVer",
+        "moduleEcoSTERSoftVer",
+        # ecoMAX850R2-X specific sensors
+        "fuelConsum",
+        "fuelStream",
+        "tempBack",
+        "transmission",
+        "statusCO",
     },
 }
 
@@ -152,10 +190,34 @@ BINARY_SENSOR_MAP_KEY = {
         "mainSrv",
         "wifi",
         "lan",
-        "lambdaStatus",
         "thermostat",
-        "statusCO",
         "statusCWU",
+        # ecoMAX850R2-X specific binary sensors
+        "contactGZC",
+        "contactGZCActive",
+        "pumpCirculationWorks",
+        "pumpSolarWorks",
+    },
+    # ecoSTER thermostat binary sensors if moduleEcoSTERSoftVer is not None
+    "ecoSter": {
+        # ecoSTER contact sensors
+        "ecoSterContacts1",
+        "ecoSterContacts2",
+        "ecoSterContacts3",
+        "ecoSterContacts4",
+        "ecoSterContacts5",
+        "ecoSterContacts6",
+        "ecoSterContacts7",
+        "ecoSterContacts8",
+        # ecoSTER day schedule sensors
+        "ecoSterDaySched1",
+        "ecoSterDaySched2",
+        "ecoSterDaySched3",
+        "ecoSterDaySched4",
+        "ecoSterDaySched5",
+        "ecoSterDaySched6",
+        "ecoSterDaySched7",
+        "ecoSterDaySched8",
     },
 }
 
@@ -212,6 +274,35 @@ ENTITY_UNIT_MAP = {
     "burnerOutput": PERCENTAGE,
     "mixerTemp": UnitOfTemperature.CELSIUS,
     "mixerSetTemp": UnitOfTemperature.CELSIUS,
+    # ecoMAX850R2-X specific units
+    "fuelConsum": PERCENTAGE,
+    "fuelStream": PERCENTAGE,
+    "transmission": None,
+    # ecoSTER thermostat units
+    "ecoSterTemp1": UnitOfTemperature.CELSIUS,
+    "ecoSterTemp2": UnitOfTemperature.CELSIUS,
+    "ecoSterTemp3": UnitOfTemperature.CELSIUS,
+    "ecoSterTemp4": UnitOfTemperature.CELSIUS,
+    "ecoSterTemp5": UnitOfTemperature.CELSIUS,
+    "ecoSterTemp6": UnitOfTemperature.CELSIUS,
+    "ecoSterTemp7": UnitOfTemperature.CELSIUS,
+    "ecoSterTemp8": UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp1": UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp2": UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp3": UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp4": UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp5": UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp6": UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp7": UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp8": UnitOfTemperature.CELSIUS,
+    "ecoSterMode1": None,
+    "ecoSterMode2": None,
+    "ecoSterMode3": None,
+    "ecoSterMode4": None,
+    "ecoSterMode5": None,
+    "ecoSterMode6": None,
+    "ecoSterMode7": None,
+    "ecoSterMode8": None,
     # ecoMAX360i
     "Circuit2thermostatTemp": UnitOfTemperature.CELSIUS,
     "TempClutch": UnitOfTemperature.CELSIUS,
@@ -230,6 +321,7 @@ STATE_CLASS_MAP: dict[str, SensorStateClass | None] = {
     "lambdaStatus": None,
     "mode": None,
     "thermostat": None,
+    "statusCO": None,
     "statusCWU": None,
     "softVer": None,
     "controllerID": None,
@@ -272,11 +364,41 @@ ENTITY_SENSOR_DEVICE_CLASS_MAP: dict[str, SensorDeviceClass | None] = {
     "mixerSetTemp": SensorDeviceClass.TEMPERATURE,
     "tempBack": SensorDeviceClass.TEMPERATURE,
     "tempCWU": SensorDeviceClass.TEMPERATURE,
+    "statusCO": None,
     "statusCWU": None,
     "tempUpperBuffer": SensorDeviceClass.TEMPERATURE,
     "tempLowerBuffer": SensorDeviceClass.TEMPERATURE,
     "signal": SensorDeviceClass.SIGNAL_STRENGTH,
     "servoMixer1": SensorDeviceClass.ENUM,
+    # ecoMAX850R2-X specific device classes
+    "fuelConsum": SensorDeviceClass.POWER_FACTOR,
+    "fuelStream": SensorDeviceClass.POWER_FACTOR,
+    "transmission": None,
+    # ecoSTER thermostat device classes
+    "ecoSterTemp1": SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp2": SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp3": SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp4": SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp5": SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp6": SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp7": SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp8": SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp1": SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp2": SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp3": SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp4": SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp5": SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp6": SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp7": SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp8": SensorDeviceClass.TEMPERATURE,
+    "ecoSterMode1": None,
+    "ecoSterMode2": None,
+    "ecoSterMode3": None,
+    "ecoSterMode4": None,
+    "ecoSterMode5": None,
+    "ecoSterMode6": None,
+    "ecoSterMode7": None,
+    "ecoSterMode8": None,
     # ecoMAX360i
     "Circuit2thermostatTemp": SensorDeviceClass.TEMPERATURE,
     "TempClutch": SensorDeviceClass.TEMPERATURE,
@@ -296,6 +418,8 @@ ENTITY_NUMBER_SENSOR_DEVICE_CLASS_MAP = {
     #############################
     "tempCOSet": NumberDeviceClass.TEMPERATURE,
     "tempCWUSet": NumberDeviceClass.TEMPERATURE,
+    # ecoMAX850R2-X specific number entities
+    "mixerSetTemp5": NumberDeviceClass.TEMPERATURE,
 }
 
 
@@ -309,8 +433,10 @@ ENTITY_BINARY_DEVICE_CLASS_MAP = {
     "lan": BinarySensorDeviceClass.CONNECTIVITY,
     "lambdaStatus": BinarySensorDeviceClass.RUNNING,
     "thermostat": BinarySensorDeviceClass.RUNNING,
-    "statusCO": BinarySensorDeviceClass.RUNNING,
     "statusCWU": BinarySensorDeviceClass.RUNNING,
+    # ecoMAX850R2-X specific binary sensors
+    "contactGZC": BinarySensorDeviceClass.CONNECTIVITY,
+    "contactGZCActive": BinarySensorDeviceClass.CONNECTIVITY,
 }
 
 """Add only keys where precision more than 0 needed"""
@@ -332,6 +458,7 @@ ENTITY_PRECISION = {
     "tempCWU": 1,
     "tempFlueGas": 1,
     "fanPower": 0,
+    "statusCO": None,
     "statusCWU": None,
     "thermostat": None,
     "lambdaStatus": None,
@@ -346,6 +473,23 @@ ENTITY_PRECISION = {
     # ecoMAX360i
     "PS": None,
     "TempBuforDown": 1,
+    # ecoSTER thermostat precision
+    "ecoSterTemp1": 1,
+    "ecoSterTemp2": 1,
+    "ecoSterTemp3": 1,
+    "ecoSterTemp4": 1,
+    "ecoSterTemp5": 1,
+    "ecoSterTemp6": 1,
+    "ecoSterTemp7": 1,
+    "ecoSterTemp8": 1,
+    "ecoSterSetTemp1": 1,
+    "ecoSterSetTemp2": 1,
+    "ecoSterSetTemp3": 1,
+    "ecoSterSetTemp4": 1,
+    "ecoSterSetTemp5": 1,
+    "ecoSterSetTemp6": 1,
+    "ecoSterSetTemp7": 1,
+    "ecoSterSetTemp8": 1,
     "heatingUpperTemp": 1,
     "Circuit1thermostat": 1,
     "heating_work_state_pump4": None,
@@ -355,6 +499,7 @@ ENTITY_ICON = {
     "mode": "mdi:sync",
     "fanPower": "mdi:fan",
     "temCO": "mdi:thermometer-lines",
+    "statusCO": "mdi:heating",
     "statusCWU": "mdi:water-boiler",
     "thermostat": "mdi:thermostat",
     "boilerPower": "mdi:gauge",
@@ -404,10 +549,70 @@ ENTITY_ICON = {
     "moduleCSoftVer": "mdi:raspberry-pi",
     "moduleLambdaSoftVer": "mdi:raspberry-pi",
     "modulePanelSoftVer": "mdi:alarm-panel-outline",
+    "moduleEcoSTERSoftVer": "mdi:raspberry-pi",
+    # ecoSTER thermostat icons
+    "ecoSterTemp1": "mdi:thermometer",
+    "ecoSterTemp2": "mdi:thermometer",
+    "ecoSterTemp3": "mdi:thermometer",
+    "ecoSterTemp4": "mdi:thermometer",
+    "ecoSterTemp5": "mdi:thermometer",
+    "ecoSterTemp6": "mdi:thermometer",
+    "ecoSterTemp7": "mdi:thermometer",
+    "ecoSterTemp8": "mdi:thermometer",
+    "ecoSterSetTemp1": "mdi:thermometer",
+    "ecoSterSetTemp2": "mdi:thermometer",
+    "ecoSterSetTemp3": "mdi:thermometer",
+    "ecoSterSetTemp4": "mdi:thermometer",
+    "ecoSterSetTemp5": "mdi:thermometer",
+    "ecoSterSetTemp6": "mdi:thermometer",
+    "ecoSterSetTemp7": "mdi:thermometer",
+    "ecoSterSetTemp8": "mdi:thermometer",
+    "ecoSterMode1": "mdi:thermostat",
+    "ecoSterMode2": "mdi:thermostat",
+    "ecoSterMode3": "mdi:thermostat",
+    "ecoSterMode4": "mdi:thermostat",
+    "ecoSterMode5": "mdi:thermostat",
+    "ecoSterMode6": "mdi:thermostat",
+    "ecoSterMode7": "mdi:thermostat",
+    "ecoSterMode8": "mdi:thermostat",
+    "ecoSterContacts1": "mdi:thermostat",
+    "ecoSterContacts2": "mdi:thermostat",
+    "ecoSterContacts3": "mdi:thermostat",
+    "ecoSterContacts4": "mdi:thermostat",
+    "ecoSterContacts5": "mdi:thermostat",
+    "ecoSterContacts6": "mdi:thermostat",
+    "ecoSterContacts7": "mdi:thermostat",
+    "ecoSterContacts8": "mdi:thermostat",
+    "ecoSterDaySched1": "mdi:calendar-clock",
+    "ecoSterDaySched2": "mdi:calendar-clock",
+    "ecoSterDaySched3": "mdi:calendar-clock",
+    "ecoSterDaySched4": "mdi:calendar-clock",
+    "ecoSterDaySched5": "mdi:calendar-clock",
+    "ecoSterDaySched6": "mdi:calendar-clock",
+    "ecoSterDaySched7": "mdi:calendar-clock",
+    "ecoSterDaySched8": "mdi:calendar-clock",
     # ecoMAX360i
+    "PS": "mdi:power-plug",
+    "Circuit2thermostatTemp": "mdi:thermometer",
+    "TempClutch": "mdi:thermometer",
+    "Circuit3thermostatTemp": "mdi:thermometer",
+    "TempWthr": "mdi:thermometer",
+    "TempCircuit3": "mdi:thermometer",
+    "TempCircuit2": "mdi:thermometer",
+    "TempBuforUp": "mdi:thermometer",
     "TempBuforDown": "mdi:thermometer",
     "heatingUpperTemp": "mdi:thermometer",
+    "Circuit1thermostat": "mdi:thermometer",
     "heating_work_state_pump4": "mdi:sync",
+    # ecoMAX850R2-X specific icons
+    "fuelConsum": "mdi:gas-station",
+    "fuelStream": "mdi:gas-station",
+    "tempBack": "mdi:thermometer",
+    "transmission": "mdi:transmission-tower",
+    "contactGZC": "mdi:connection",
+    "contactGZCActive": "mdi:connection",
+    "pumpCirculationWorks": "mdi:pump",
+    "pumpSolarWorks": "mdi:pump",
 }
 
 ENTITY_ICON_OFF = {
@@ -426,9 +631,29 @@ ENTITY_ICON_OFF = {
     "mainSrv": "mdi:server-network-off",
     "lan": "mdi:lan-disconnect",
     "lighterWorks": "mdi:fire-off",
-    "lambdaStatus": "mdi:lambda-off",
     "thermostat": "mdi:thermostat-off",
-    "statusCO": "mdi:heating-off",
+    # ecoMAX850R2-X specific off icons
+    "contactGZC": "mdi:connection-off",
+    "contactGZCActive": "mdi:connection-off",
+    "pumpCirculationWorks": "mdi:pump-off",
+    "pumpSolarWorks": "mdi:pump-off",
+    # ecoSTER thermostat off icons
+    "ecoSterContacts1": "mdi:thermostat-off",
+    "ecoSterContacts2": "mdi:thermostat-off",
+    "ecoSterContacts3": "mdi:thermostat-off",
+    "ecoSterContacts4": "mdi:thermostat-off",
+    "ecoSterContacts5": "mdi:thermostat-off",
+    "ecoSterContacts6": "mdi:thermostat-off",
+    "ecoSterContacts7": "mdi:thermostat-off",
+    "ecoSterContacts8": "mdi:thermostat-off",
+    "ecoSterDaySched1": "mdi:calendar-clock-off",
+    "ecoSterDaySched2": "mdi:calendar-clock-off",
+    "ecoSterDaySched3": "mdi:calendar-clock-off",
+    "ecoSterDaySched4": "mdi:calendar-clock-off",
+    "ecoSterDaySched5": "mdi:calendar-clock-off",
+    "ecoSterDaySched6": "mdi:calendar-clock-off",
+    "ecoSterDaySched7": "mdi:calendar-clock-off",
+    "ecoSterDaySched8": "mdi:calendar-clock-off",
 }
 
 NO_CWU_TEMP_SET_STATUS_CODE = 128