+++ custom_components/econet300/entity.py
@@ -10,6 +10,7 @@ from .api import Econet300Api
 from .common import EconetDataCoordinator
 from .const import (
     DEVICE_INFO_CONTROLLER_NAME,
+    DEVICE_INFO_ECOSTER_NAME,
     DEVICE_INFO_LAMBDA_NAME,
     DEVICE_INFO_MANUFACTURER,
     DEVICE_INFO_MIXER_NAME,
@@ -204,6 +205,37 @@ class LambdaEntity(EconetEntity):
             name=f"{DEVICE_INFO_LAMBDA_NAME}",
             manufacturer=DEVICE_INFO_MANUFACTURER,
             model=DEVICE_INFO_MODEL,
+            configuration_url=self.api.host,
+            sw_version=self.api.sw_rev,
+            via_device=(DOMAIN, self.api.uid),
+        )
+
+
+class EcoSterEntity(EconetEntity):
+    """Represents EcoSterEntity."""
+
+    def __init__(
+        self,
+        description: EntityDescription,
+        coordinator: EconetDataCoordinator,
+        api: Econet300Api,
+        idx: int,
+    ):
+        """Initialize the EcoSterEntity."""
+        self.entity_description = description
+        self.api = api
+        self._idx = idx
+        super().__init__(coordinator)
+
+    @property
+    def device_info(self) -> DeviceInfo | None:
+        """Return device info of the entity."""
+        return DeviceInfo(
+            identifiers={(DOMAIN, f"{self.api.uid}-ecoster-{self._idx}")},
+            name=f"{DEVICE_INFO_ECOSTER_NAME} {self._idx}",
+            manufacturer=DEVICE_INFO_MANUFACTURER,
+            model=DEVICE_INFO_MODEL,
+            model_id=self.api.model_id,
             configuration_url=self.api.host,
             sw_version=self.api.sw_rev,
             via_device=(DOMAIN, self.api.uid),